chore: закрыть все замечания ruff и basedpyright
Ruff: 19× I001 (порядок импортов в тестах и src) — autofix.

basedpyright (58 → 0):
- pyrightconfig.json: tests-окружению добавлен '.' в extraPaths —
  'tests.conftest' больше не unresolved (31 error).
- src/json_viewer.py: убран FakeEvent (вызов on_tree_node_selected с
  самодельным событием); общий путь _handle_node_selected(node) +
  node/app через Any/геттеры — hasattr-гварды сохранены, поведение то же.
- src/kctx_store.py: lock_timeout аннотирован int (как у acquire_file_lock).
- Тесты: query(CommandBlock) вместо строкового селектора (raw_stdout/
  header/line_nav_active на Widget), сужение line_index/Path|None перед
  сравнениями и load_scenario, evaluate(...) перед 'in', str(style),
  cast(Any,{}) для HTTPError, getattr(handler,'proxies',{}).

Проверка: ruff 0; basedpyright 0 errors/0 warnings/0 notes;
202 passed (затронутые наборы: ipcalc, llm, screensaver, update_check,
tags, demo, json_viewer, kctx_*, completion).

+++ src/kctx_store.py
@@ -143,7 +143,7 @@ def add_snapshot(
     now: float | None = None,
     per_cluster_limit: int = DEFAULT_PER_CLUSTER_LIMIT,
     total_limit: int = DEFAULT_TOTAL_LIMIT,
-    lock_timeout: float = DEFAULT_LOCK_TIMEOUT,
+    lock_timeout: int = DEFAULT_LOCK_TIMEOUT,
 ) -> bool:
     """Сохраняет снимок kubectl-стека для кластера.
